refactor, cleanup style structure
