--- libs/ui/src/components/atoms/tooltip/test.ts
@@ -1,3 +1,0 @@
-import { defaultDarkTheme } from '../../../theme';
-
-export const test = defaultDarkTheme;
